CrashHandler improvements

+++ malheur/src/main/java/de/kuschku/malheur/data/AppInfo.kt
@@ -21,6 +21,6 @@ package de.kuschku.malheur.data
 
 data class AppInfo(
   val versionName: String?,
-  val versionCode: Int?,
+  val versionCode: Long?,
   val buildConfig: Map<String, Any?>?
 )
